added Advance DS and Algo. project
PROBLEM STATEMENT

ALTERNATING DISKS:
You have a row 2n disks of two colors , n dark and n light. They alternate : dark , light and so on. You want to get all
the dark disks to the right end and all the lights disks to the left-hand end. The only moves you are allowed to make are
those that interchange the positons of two neighbouring disks.